Add file headers and cleanup structure
